+++ sdk/browser/src/AffinityWallet.ts
@@ -1,9 +1,7 @@
-import { KeysService } from '@affinidi/common'
-import { CommonNetworkMember as CoreNetwork, __dangerous } from '@affinidi/wallet-core-sdk'
 import { EventComponent } from '@affinidi/affinity-metrics-lib'
-
-import WalletStorageService from './services/WalletStorageService'
+import { CommonNetworkMember as CoreNetwork, __dangerous } from '@affinidi/wallet-core-sdk'
 import { FetchCredentialsPaginationOptions } from '@affinidi/wallet-core-sdk/dist/dto/shared.dto'
+
 import platformEncryptionTools from './PlatformEncryptionTools'
 
 export type SdkOptions = __dangerous.SdkOptions & {
@@ -13,21 +11,13 @@ export type SdkOptions = __dangerous.SdkOptions & {
 const COMPONENT = EventComponent.AffinidiBrowserSDK
 
 export class AffinityWallet extends CoreNetwork {
-  keysService: KeysService
-  walletStorageService: WalletStorageService
-
   constructor(
     password: string,
     encryptedSeed: string,
     options: __dangerous.SdkOptions = {},
     component: EventComponent = COMPONENT,
   ) {
-    super(password, encryptedSeed, options, component)
-
-    const sdkOptions = CoreNetwork.setEnvironmentVarialbles(options)
-
-    this.keysService = new KeysService(encryptedSeed, password)
-    this.walletStorageService = new WalletStorageService(encryptedSeed, password, sdkOptions)
+    super(password, encryptedSeed, platformEncryptionTools, options, component)
   }
 
   /**
@@ -44,8 +34,8 @@ export class AffinityWallet extends CoreNetwork {
     const { keyStorageUrl, userPoolId } = CoreNetwork.setEnvironmentVarialbles(options)
     const { accessToken } = __dangerous.readUserTokensFromSessionStorage(userPoolId)
 
-    const encryptedSeed = await WalletStorageService.pullEncryptedSeed(accessToken, keyStorageUrl, options)
-    const encryptionKey = await WalletStorageService.pullEncryptionKey(accessToken)
+    const encryptedSeed = await __dangerous.WalletStorageService.pullEncryptedSeed(accessToken, keyStorageUrl, options)
+    const encryptionKey = await __dangerous.WalletStorageService.pullEncryptionKey(accessToken)
 
     return new AffinityWallet(encryptionKey, encryptedSeed, options)
   }
@@ -94,7 +84,7 @@ export class AffinityWallet extends CoreNetwork {
    * @returns decrypted message
    */
   async readEncryptedMessage(encryptedMessage: string): Promise<any> {
-    const privateKeyBuffer = this.keysService.getOwnPrivateKey()
+    const privateKeyBuffer = this._keysService.getOwnPrivateKey()
 
     return platformEncryptionTools.decryptByPrivateKey(privateKeyBuffer, encryptedMessage)
   }
@@ -108,7 +98,7 @@ export class AffinityWallet extends CoreNetwork {
    * @returns array of ids for corelated records
    */
   async saveCredentials(data: any[], storageRegion?: string): Promise<any> {
-    const result = await this.walletStorageService.saveUnencryptedCredentials(data, storageRegion)
+    const result = await this._walletStorageService.saveUnencryptedCredentials(data, storageRegion)
 
     this._sendVCSavedMetrics(data)
     // NOTE: what if creds actually were not saved in the vault?
@@ -124,7 +114,7 @@ export class AffinityWallet extends CoreNetwork {
    */
   async getCredentialByIndex(credentialIndex: number): Promise<any> {
     const paginationOptions: FetchCredentialsPaginationOptions = { skip: credentialIndex, limit: 1 }
-    const credentials = await this.walletStorageService.fetchDecryptedCredentials(paginationOptions)
+    const credentials = await this._walletStorageService.fetchDecryptedCredentials(paginationOptions)
 
     if (!credentials[0]) {
       throw new __dangerous.SdkError('COR-14')
@@ -143,10 +133,10 @@ export class AffinityWallet extends CoreNetwork {
    * @returns array of VCs
    */
   async getCredentials(credentialShareRequestToken: string = null): Promise<any[]> {
-    const allCredentials = await this.walletStorageService.fetchAllDecryptedCredentials()
+    const allCredentials = await this._walletStorageService.fetchAllDecryptedCredentials()
 
     if (credentialShareRequestToken) {
-      return this.walletStorageService.filterCredentials(credentialShareRequestToken, allCredentials)
+      return this._walletStorageService.filterCredentials(credentialShareRequestToken, allCredentials)
     }
 
     return allCredentials
@@ -164,7 +154,7 @@ export class AffinityWallet extends CoreNetwork {
       })
     }
 
-    const credentialIndexToDelete = credentialIndex ?? (await this.walletStorageService.findCredentialIndexById(id))
+    const credentialIndexToDelete = credentialIndex ?? (await this._walletStorageService.findCredentialIndexById(id))
     return this.deleteCredentialByIndex(credentialIndexToDelete.toString())
   }
 }