+++ sdk/core/test/helpers/CommonNetworkMember.ts
@@ -1,10 +1,11 @@
 import { EventComponent } from '@affinidi/affinity-metrics-lib'
 import { CommonNetworkMember as GenericCommonNetworkMember } from '../../src/CommonNetworkMember'
 import { SdkOptions } from '../../src/dto/shared.dto'
+import { testPlatformTools } from './testPlatformTools'
 
 export class CommonNetworkMember extends GenericCommonNetworkMember {
   constructor(password: string, encryptedSeed: string, options: SdkOptions = {}, component?: EventComponent) {
-    super(password, encryptedSeed, options, component)
+    super(password, encryptedSeed, testPlatformTools, options, component)
   }
 
   static async afterConfirmSignUp() {