+++ sdk/expo/CHANGELOG.md
@@ -8,7 +8,7 @@
   * `getCredentials` no longer accepts pagination parameters
   * Some methods that returned arrays now declare their return type as `any[]` instead of `any`.
   * Some methods that accepted arrays now accept `unknown[]` instead of `any`.
-  * Removed `KeysService`
+  * Removed `KeysService` and `WalletStorageService`
 # release 4.3.1 (2021-06-07)
 * Fixed DID validation
 # release 4.3.0 (2021-06-04)